Add Initial new uni types changes

+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v1/x0/x0/datatypes/CreateCommentLike.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 internal data class CreateCommentLike(
     val comment_id: CommentId,
-    val score: Long,
+    val score: Int,
 )
